Refactoring and preparing for pull request
