+++ src/components/ConditionSelector/index.ts
@@ -3,3 +3,4 @@ export * from './Condition.Props';
 export * from './ConditionGroup';
 export * from './ConditionGroup.Props';
 export * from './ConditionSelector';
+export * from './ConditionSelector.Props';
